fix: resolve 10 backend bugs — delete validation, property mismatches, missing model fields
- TPI: Add DeleteValidationHelper check in RemoveTPI (was missing unlike Bank)
- Supplier: Add BlacklistDate to GetAllSuppliers projection
- Supplier: Return NotFound instead of NoContent when entity is null
- DimensionalFactor: Handle null/0 for ParameterUnitID and DefaultTestMethodID
- TestMaster: Add RemoveTestMaster call + fix message "created" → "deleted"
- ProductTestGroup: Rename TestMethodStandardName → TestMethodSpecificationName
- ProductSpecificationGrade: Rename GradeName → SpecificationGradeName
- GstConfig: Restore DefaultGstRate and PIGstApplicable (removed by stash not applied)
- GstDto: Restore PIGstApplicable property
- SettingsService: Add decimal→int cast for DefaultGstRate

+++ LIMSApi/ServiceWORepo/SettingsService.cs
@@ -286,7 +286,7 @@ namespace LIMSApi.ServiceWORepo
                 Gstin = gst?.GstNumber,
                 PanNumber = null,
                 StateCode = gst?.State,
-                DefaultGstRate = gst?.DefaultGstRate ?? 18,
+                DefaultGstRate = (int)(gst?.DefaultGstRate ?? 18),
                 PIGstApplicable = gst?.PIGstApplicable ?? true
             };
